move dashboard component tests to adhere to jest convention
