+++ dashboard-app/src/dashboard/ByTime/__tests__/ByTime.test.ts
@@ -1,8 +1,6 @@
-import {
-  cleanup,
-} from 'react-testing-library';
-import { renderWithHistory } from '../../util/tests';
-import ByTime from '../ByTime';
+import { cleanup } from 'react-testing-library';
+import { renderWithHistory } from '../../../util/tests';
+import ByTime from '..';
 import 'jest-dom/extend-expect';
 
 describe('By Activity Page', () => {
